cloud_deployment/functions/process_batches/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.21.1 